% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataEB.R
\docType{data}
\name{dataEB}
\alias{dataEB}
\title{Sample Data for Practice}
\format{
A sample data has 3 column, which are:
\describe{
\item{code}{code of each area}
\item{y}{status "success" or not in each unit sample of each area}
\item{weight}{a weight value in each unit sample of each area}
}
}
\usage{
dataEB
}
\description{
An example data for trying and testing in saebnocov package
}
\examples{
data(dataEB)
}
\keyword{datasets}
